Assignment in hibernate

+++ src/main/java/com/example/complexhibernate/assignment/models/Friend.java
@@ -3,6 +3,7 @@ package com.example.complexhibernate.assignment.models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -15,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Entity
 public class Friend {
     @Id
-    private int id;
+    private Long id;
     private String namn;
     private String adress;
     private String mobilNr;
